refactor(pension-plans): apply project guidelines to parameters, imports, and error handling
